Suppress pack errors in balloon, vioserial, vioinput
A few harmless errors are printed when building balloon,
vioserial, and vioinput. This commit fixes two issues:

* Win10 drivers don't ship with WDF coninstallers so
  WdfCoInstaller*.dll should not be copied over to Install.

* The Install directory may be empty or not exist, which
  is worth checking before running del on its content.

+++ Balloon/sys/packOne.bat
@@ -43,15 +43,15 @@ set INF_PATH_AND_NAME=objfre_%INST_OS%_%INST_ARC%\%INST_EXT%\%FILE_NAME%.inf
 set WDF_PATH_AND_NAME=%BUILDROOT%\redist\wdf\%INST_ARC%\WdfCoInstaller01009.dll
 if /i "%1"=="win7" set WDF_PATH_AND_NAME="C:\Program Files (x86)\Windows Kits\10\Redist\wdf\%2\WdfCoInstaller01009.dll"
 if /i "%1"=="win8" set WDF_PATH_AND_NAME="C:\Program Files (x86)\Windows Kits\10\Redist\wdf\%2\WdfCoInstaller01011.dll"
-if /i "%1"=="win10" set WDF_PATH_AND_NAME="C:\Program Files (x86)\Windows Kits\10\Redist\wdf\%2\WdfCoInstaller01015.dll"
+if /i "%1"=="win10" set WDF_PATH_AND_NAME=
 
 echo makeinstall %1 %2 %3
 mkdir ..\Install\%INST_OS%\%INST_ARC%
-del /Q ..\Install\%INST_OS%\%INST_ARC%\%FILE_NAME%.*
+if exist ..\Install\%INST_OS%\%INST_ARC%\%FILE_NAME%.* del /Q ..\Install\%INST_OS%\%INST_ARC%\%FILE_NAME%.*
 copy /Y %SYS_PATH_AND_NAME% ..\Install\%INST_OS%\%INST_ARC%
 copy /Y %PDB_PATH_AND_NAME% ..\Install\%INST_OS%\%INST_ARC%
 copy /Y %INF_PATH_AND_NAME% ..\Install\%INST_OS%\%INST_ARC%
-copy /Y %WDF_PATH_AND_NAME% ..\Install\%INST_OS%\%INST_ARC%
+if not [%WDF_PATH_AND_NAME%] == [] copy /Y %WDF_PATH_AND_NAME% ..\Install\%INST_OS%\%INST_ARC%
 
 :create_cat
 echo "Setting OS mask for:" %1 %2
